fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@20669 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSLeaderInfo.cpp
@@ -142,7 +142,7 @@ MSLeaderInfo::getSubLanes(const MSVehicle* veh, SUMOReal latOffset, int& rightmo
 const MSVehicle*
 MSLeaderInfo::operator[](int sublane) const {
     assert(sublane >= 0);
-    assert(sublane < myVehicles.size());
+    assert(sublane < (int)myVehicles.size());
     return myVehicles[sublane];
 }
 
@@ -241,7 +241,7 @@ MSLeaderDistanceInfo::clear() {
 CLeaderDist
 MSLeaderDistanceInfo::operator[](int sublane) const {
     assert(sublane >= 0);
-    assert(sublane < myVehicles.size());
+    assert(sublane < (int)myVehicles.size());
     return std::make_pair(myVehicles[sublane], myDistances[sublane]);
 }
